Move app_widget.dart in his own file inside presentation
